快速生成option 模板
